Fix: Static Libraries and Zero Warnings

+++ Hazel/src/Platform/Windows/WindowsWindow.cpp
@@ -128,13 +128,13 @@ namespace Hazel {
 
 		glfwSetCursorPosCallback(m_Window, [](GLFWwindow* window, double xpos, double ypos) {
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-			MouseMoveEvent event(xpos, ypos);
+			MouseMoveEvent event((float)xpos, (float)ypos);
 			data.EventCallback(event);
 		});
 
 		glfwSetScrollCallback(m_Window, [](GLFWwindow* window, double xoffset, double yoffset) {
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-			MouseScrolledEvent event(xoffset, yoffset);
+			MouseScrolledEvent event((float)xoffset, (float)yoffset);
 			data.EventCallback(event);
 		});
 	}
